Règle d'écrêtement pour MPRG
Fixes #38
